scripts/video_gen_tests/video_agent/requirements.txt
claude-agent-sdk>=0.1.50
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
google-genai>=1.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pydantic>=2.10.0
websockets>=14.0
python-multipart>=0.0.20
